Development
DirectoriesHandler node and script were created to handle files and directories.
The DirectoriesHandler script was modified to (1) initialize the system documents path by default, (2) get and set the working directory, (3) check if a directory or file exists, and (4) check if a directory is empty.

Controller node and script were created to control the core of the program.
DirectoriesHandler node was added as a child of the Controller node.
The Controller script was modified to (1) run a complete test of the DirectoriesHandler script implementations.