Refactoring
The directory structure of the script files was modified to allow future implementations of Test Driven Development.

DirectoriesHandler node and script were renamed as PathsHandler respectively to improve the understanding of their porpuse.
The structure of the functions of the PathsHandler script was modified to improve its functionalities.

Controller script was modified, deleting all its content to allow future implementations of Test Driven Development in an easy way.