Refactoring
TestersHandler script was created to handle all the testers scripts of the program in a simple way.
The TestersHandler script was modified to (1) provide a function that allows the comparison between a resulting value and an expected value, (2) add testers scripts as children of the script, and (3) provide a function that runs all the test of the testers scripts.

PathsHandlerTester script was created to run individual tests for each implementation of the PathsHandler script.
The PathsHandlerTester script was modified (1) get the names of the test functions, (2) provide a function to run the test functions in a sequential way, and (3) cover all tests performed so far.

The Controller script was modified to (1) allow the implementation of Test Driven Development in an automatic way.