Development
CommandsExecuter node and script were created to execute OS commands to work with Blender engine.
The CommandsExecuter script was modified to (1) check the functionality of the Blender engine path.

CommandsExecuterTester script was created to run individual tests for each implementation of the CommandsExecuter script.

The PathsHandler script was modified to (1) improve its description.

The TestersHandler script was modified to add the CommandsExecuterTester script as a child.