Development
RendersScaler node and script were created to scale PNG images.
The RendersScaler script was modified to (1) get images dimensions, (2) check scaling possibilities, and (3) scale PNG images.

RendersScalerTester script was created to run individual tests for each implementation of the RendersScaler script.

The TestersHandler script was modified to add the RendersScalerTester script as a child.